// src/buildpc/interfaces/buildpc.ts
export interface BuildPC {
    mothers?:     Mother[];
    cpu?:         CPU[];
    coolercpu?:   Coolercpu[];
    mem?:         Mem[];
    video?:       Video[];
    hd?:          HD[];
    gab?:         Gab[];
    fuente?:      Fuente[];
    monitores?:   Monitore[];
    auriculares?: Auriculare[];
    card_reader?: any[];
    coolerfan?:   Coolerfan[];
    mouse?:       Mouse[];
    mousepad?:    Mousepad[];
    opticos?:     any[];
    os?:          any[];
    soportes?:    any[];
    teclados?:    Teclado[];
    extras?:      any[];
    notebooks?:   Notebook[];
}

export interface Auriculare {
    id_producto:         number;
    disponible:          number;
    mini_plug_audio:     null | string;
    mini_plug_combinado: null | string;
    impedancia:          null | string;
    usb:                 string;
    wifi_por_receptor:   string;
    wifi_por_bt:         string;
    tamano_parlante:     string;
    tipo:                string;
    resp_hz_min:         string;
    resp_hz_max:         string;
    sensibilidad:        null | string;
    colores:             Colores;
    mini_plug_mic:       null | string;
    usb_power:           string;
    disp_compatible:     string;
    largo_cable:         null | string;
    iluminacion:         ColorLuzTecladoEnum | null;
    base:                string;
    mic_resp_hz_min:     null | string;
    mic_resp_hz_max:     null | string;
    mic_sensibilidad:    null;
    mic_tipo:            MicTipo | null;
    mic_cancelacion:     string;
    mic_tipo_captacion:  MicTipoCaptacion | null;
    surround:            string;
    wireless:            Wireless;
    con_mic:             string;
    virtual_real:        VirtualReal;
    formato:             Formato;
}

export enum Colores {
    Blanco = "Blanco",
    Negro = "Negro",
    NegroAmarillo = "Negro,Amarillo",
    NegroRojo = "Negro,Rojo",
    Rosa = "Rosa",
}

export enum Formato {
    Headset = "Headset",
}

export enum ColorLuzTecladoEnum {
    RGB = "RGB",
    Rojo = "Rojo",
    SinIluminación = "Sin Iluminación",
}
// ...
export interface Coolercpu {
    id_producto:       number;
    sockets:           string;
    disponible:        number | null;
    altura:            number;
    tipo_gab:          TipoGab;
    espacios_120mm:    string;
    espacios_140mm:    number;
    altura_real:       string;
    nro_coolers:       number;
    w:                 number;
    cooler_base:       number;
    tipo_disipacion:   TipoDisipacion | null;
    color_iluminacion: null | string;
    ruido:             null;
    ruido_max:         null | string;
    sistema_rgb:       CoolercpuSistemaRGB | null;
    TDP:               null | string;
    tam_cooler:        null | string;
}
// ...
export interface CPU {
    id_producto:         number;
    modelo:              string;
    cores:               number;
    disponible:          number;
    w:                   number;
    clock:               number;
    tipo:                string;
    cooler:              number;
    familia:             string;
    l1:                  null | string;
    l2:                  null | string;
    l3:                  null | string;
    tecnologia:          string;
    cate_cpu:            number;
    chipset_gpu:         string;
    nivel_cpu_deprecado: null;
    nivel_gpu:           number;
    hilos:               number;
    turbo:               number | null;
}

export interface Fuente {
    id_producto:      number;
    disponible:       number;
    w_real:           number;
    w_nominal:        number;
    tipo:             TipoFuenteEnum;
    marca:            null;
    _24pines:         number;
    _4pines:          number;
    _4plus:           number;
    pines_pcie:       number;
    pines_pcieplus:   number;
    incluye_cable:    number;
    kit:              number;
    compatible_abajo: number;
    certi80plus:      TresD | null;
    ModoHibrido:      number;
    tipocableado:     Tipocableado;
    amper12v:         number | null;
    sata:             number;
    molex:            string;
    floppy:           null;
    digital:          number;
    color:            Color;
    iluminacion:      ColorLuzTecladoEnum | null;
}
// ...
export interface Gab {
    id_producto:                        number;
    kit:                                number;
    kit_detalle:                        null | string;
    disponible:                         number;
    w:                                  number;
    tipo_fuente:                        TipoFuenteEnum;
    id_fuente:                          number;
    posicion_fuente_superior:           number;
    tipo_mother:                        TipoMother;
    altura_cpu:                         number;
    unidades_opticas:                   number;
    cf_80mm:                            number;
    cf_120mm:                           number;
    cf_140mm:                           number;
    cf_200mm:                           number;
    cf_80mm_incluidos:                  number;
    cf_120mm_incluidos:                 number;
    cf_140mm_incluidos:                 number;
    cf_200mm_incluidos:                 number;
    cf_2x120mm_continuos:               number;
    cf_2x140mm_continuos:               number;
    cf_3x120mm_continuos:               number;
    cf_3x140mm_continuos:               number;
    cf_continuos_con_espacio_disipador: number;
    slots:                              number;
    ancho_maximo_vga:                   number;
    unidades_2_5:                       number;
    unidades_3_5:                       number;
    unidades_3_25:                      number;
    ancho:                              number | null;
    alto:                               number | null;
    profundidad:                        number | null;
    largo_maximo_vga:                   number;
    usb30_frontal:                      number;
    usb20_frontal:                      number;
    usb31_frontal:                      number;
    conector_usb30_interno:             number;
    conector_usb20_interno:             number;
    conector_usb31_interno:             number;
    audio_ac97_frontal:                 number;
    audio_hd_frontal:                   number;
    lector_tarjetas_frontal:            number;
    ventana:                            number;
    tipo_de_ventana:                    TipoDeVentana | null;
    iluminacion:                        GabIluminacion | null;
    color:                              Color;
    Soporte_GPU_Vertical:               number;
    riser:                              number;
    controlador_de_fans:                ControladorDeFans | null;
    Controlador_RGB:                    number;
    Cover_fuente:                       number;
    altura_ccpu:                        number;
    sistema_rgb:                        CoolercpuSistemaRGB | null;
    tipo_ponderado:                     null;
    factor:                             TipoGab;
}
// ...
export interface HD {
    id_producto:                    number;
    marca:                          null | string;
    capacidad:                      number;
    tasa_transf:                    number;
    disponible:                     number;
    cache:                          number;
    rpm:                            number;
    tipo:                           HDTipo;
    w:                              number;
    dimension:                      Dimension;
    tipo_almacenamiento:            TipoAlmacenamiento;
    formato_notebook:               number;
    velocidad_lectura_secuencial:   number | null;
    velocidad_escritura_secuencial: number | null;
    velocidad_lectura_random:       number | null;
    velocidad_escritura_random:     number | null;
    vida_util:                      number | null;
    tbw:                            number | null;
    categoria_tamaño:               CategoriaTamaño;
    prod_hdcol:                     null;
    interfaz:                       Interfaz;
}
// ...
export interface Mem {
    id_producto:     number;
    capacidad:       number;
    disponible:      number;
    disipador:       number;
    velocidad:       string;
    tipo:            PurpleTipo;
    w:               number;
    slot:            number;
    sodimm:          number;
    activo:          number;
    latencia_cl:     number;
    voltaje:         number;
    marca:           string;
    disipacion_alta: number;
}

export enum PurpleTipo {
    Ddr4 = "DDR4",
    Ddr5 = "DDR5",
}

export interface Monitore {
    id_producto:               number;
    disponible:                number;
    pulgadas:                  number;
    pulgadas_visibles:         number;
    resolucion:                Resolucion;
    max_frecuenta:             number;
    radio_constraste_estativo: number | null;
    radio_constraste_dynamico: number | null;
    tipo_de_ilumunacion:       null;
    brillo:                    number | null;
    colores:                   number | null;
    tiempo_respuesta:          number;
    angulo_vision_horizontal:  number;
    angulo_vision_vertical:    number;
    vida_backlight:            number | null;
    tipo_de_panel:             string;
    entra_3_5mm:               number;
    salida_3_5mm:              number;
    HDMI:                      number;
    DVI:                       null;
    VGA:                       number;
    DP:                        number;
    cant_parlantes:            number;
    potencia_x_parlante:       number | null;
    ConsumoMax:                number | null;
    VESA:                      VESA;
    peso:                      number;
    ancho:                     number;
    alto:                      number;
    espesor:                   number;
    sintonizador_analogico:    number;
    sintonizador_tda:          number;
    USB20:                     number;
    USB30:                     number;
    USB31:                     number;
    smart:                     number;
    control_remoto:            number;
    android_tv:                number;
    pie_desmontable:           number;
    cables:                    null | string;
    tipo_fuente:               TipoFuente;
    nvidia_g_sync:             number;
    amd_freesync:              number;
    curvo:                     number;
    curvatura:                 number;
    tres_d:                    TresD;
    tactil:                    number;
    mini_dp:                   number;
    display_wifi:              number;
    entrada_3_5_mic:           number;
}
// ...
export interface Mother {
    id_producto:     number;
    modelo:          null | string;
    cpus:            string;
    SATA:            number;
    marca:           null | string;
    _4pines:         number;
    _4plus:          number;
    wmax:            number;
    _24pines:        number;
    PCI:             number;
    nro_pcie:        number;
    PCIE1X:          number;
    nro_agp:         number;
    IDE:             number;
    tipo_multiplaca: string;
    DDR:             number;
    DDR2:            number;
    DDR3:            number;
    DDR4:            number;
    DDR5:            number;
    Chipsets:        string;
    disponible:      number;
    vga:             number;
    HDMI:            number;
    dvi:             string;
    DP:              number;
    w:               number;
    Floppy:          number;
    USB:             number;
    Factor:          TipoGab;
    sodimm:          number;
    m2:              number;
    con_micro:       number;
    id_sub_cate:     number;
    tipo:            TipoCPUEnum;
    activo:          number;
    wifi_integrada:  number;
    sistema_rgb:     MotherSistemaRGB | null;
    usb20_trasero:   number;
    usb30_trasero:   number;
    usb31_trasero:   number;
    ps2:             number;
    lan:             LAN;
    placa_audio:     string;
    anclaje:         string;
    bios_flashback:  number;
    usb_32_trasero:  number;
    usb_type_c:      number;
    pci_e_x4:        number;
    m2_sata:         number;
    m2_nvme:         number;
    tipo_memoria:    PurpleTipo;
    slot_memoria:    number;
}